+++ sources/structure/subsystem.h
@@ -187,14 +187,14 @@ class integral_datum_item
     std::vector<int> diagonal; // inv.factors image $(1-\theta)X^*$ in $X^*/N$
     int_Matrix in, out; // from $X^*/N$ to $-1$ subspace to $(1-\theta)X^*$
     codec (const InnerClass& ic,
-	   unsigned int isys, InvolutionNbr inv, const int_Matrix& cmat);
+	   unsigned long int isys, InvolutionNbr inv, const int_Matrix& cmat);
   }; // |struct integral_datum_item::codec|
 
   integral_datum_item(InnerClass& ic,const RootNbrSet& int_posroots);
   integral_datum_item(integral_datum_item&&)=default; // move, never copy
 
   const SubSystem& int_system() const { return *integral; }
-  codec data(const InnerClass& ic, unsigned int isys, InvolutionNbr inv) const
+  codec data(const InnerClass& ic, unsigned long int isys, InvolutionNbr inv) const
   { return { ic,isys,inv,simple_coroots }; }
   const int_Matrix& coroots_matrix() const { return simple_coroots; }
